+++ matchbox-engine/src/test/java/ch/ahdis/matchbox/engine/tests/FhirMappingLanguageTests.java
@@ -1300,6 +1300,7 @@ class FhirMappingLanguageTests {
 				"http://test.ch/DummyBundleToBundle", true);
 		assertTrue(result != null);
 
+		CompareUtil.logMemory();
 	}
 
 }