+++ docs/changelog.md
@@ -1,8 +1,9 @@
-2025/01/xx Release 3.9.10
+2025/01/13 Release 3.9.10
 - Performance improvement fml [#323](parsing https://github.com/ahdis/matchbox/issues/323)
 - Update org.hl7.fhir.core to 6.5.4 and hapi-fhir to 7.6.1 
 - Update integration tests with correct url
 - make autoinstall new ig [#325](https://github.com/ahdis/matchbox/issues/325)
+- support for terminology servers which require authentication [#327](https://github.com/ahdis/matchbox/issues/327), thanks @echiu-infoway for support!
 
 2024/12/09 Release 3.9.9